Added colours, based on speed.
Lower speeds are displayed in red, higher speeds blend to blue. The
code now draws only the last bit of the line, instead of redrawing
everything on every frame. This means, that I had to remove the
visualization of how the patterns is generated, but at least it runs
faster now.
The code is not very nice at the moment, but it works.